+++ oshdb-api/src/main/java/org/heigit/ohsome/oshdb/api/db/OSHDBH2.java
@@ -1,10 +1,8 @@
 package org.heigit.ohsome.oshdb.api.db;
 
-import static java.util.Optional.ofNullable;
 import static org.heigit.ohsome.oshdb.api.db.H2Support.pathToUrl;
 
 import java.nio.file.Path;
-import javax.sql.DataSource;
 import org.h2.jdbcx.JdbcConnectionPool;
 
 /**
@@ -24,31 +22,42 @@ public class OSHDBH2 extends OSHDBJdbc {
     this(databaseFile, "sa", "");
   }
 
+  /**
+   * Opens a connection to oshdb data stored in a H2 database file.
+   *
+   * @param databaseFile the file name and path to the H2 database file. (the ".mv.db" file ending
+   *        of H2 should be omitted here)
+   */
   public OSHDBH2(String databaseFile) {
     this(Path.of(databaseFile));
   }
 
+  /**
+   * Opens a connection to oshdb data stored in a H2 database file.
+   *
+   * @param url the JDBC URL to the H2 database file.
+   * @param user username for the H2 JDBC connection
+   * @param password password for the H2 JDBC connection
+   */
   public OSHDBH2(String url, String user, String password) {
     this(JdbcConnectionPool.create(url, user, password));
   }
 
+  /**
+   * Opens a connection to oshdb data stored in a H2 database file.
+   *
+   * @param path the file name and path to the H2 database file. (the ".mv.db" file ending
+   *        of H2 should be omitted here)
+   * @param user username for the H2 JDBC connection
+   * @param password password for the H2 JDBC connection
+   */
   public OSHDBH2(Path path, String user, String password) {
     this(pathToUrl(path), user, password);
   }
 
-  public OSHDBH2(DataSource ds) {
-    super(ds);
-    this.connectionPool = null;
-  }
-
   private OSHDBH2(JdbcConnectionPool ds) {
-    super(ds);
+    super(ds, "");
     this.connectionPool = ds;
-  }
-
-  @Override
-  public OSHDBH2 prefix(String prefix) {
-    return (OSHDBH2) super.prefix(prefix);
   }
 
   @Override
@@ -59,7 +68,7 @@ public class OSHDBH2 extends OSHDBJdbc {
   @Override
   public void close() throws Exception {
     try {
-      ofNullable(connectionPool).ifPresent(JdbcConnectionPool::dispose);
+      connectionPool.dispose();
     } finally {
       super.close();
     }