+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog
 * remove deprecated filter methods `osmType`, `osmTag` and `osmEntityFlter` (which were replaced by [OSHDB `filter`s](https://github.com/GIScience/oshdb/blob/0.7/documentation/manual/filters.md#oshdb-filter)) ([#451])
 * `OSMType.toString` returns lower-case name. ([#459])
 * rework of handling separate keytables, see [#470] ([#470])
+* `prefix` is now a mandatory parameter for OSHDB(Jdbc|Ignite). ([#478])
+* `OSHDBH2` no longer supports `DataSource`s as argument. ([#478])
 
 ### new features
 
@@ -50,6 +52,7 @@ Changelog
 [#459]: https://github.com/GIScience/oshdb/pull/459
 [#467]: https://github.com/GIScience/oshdb/pull/467
 [#470]: https://github.com/GIScience/oshdb/pull/470
+[#478]: https://github.com/GIScience/oshdb/pull/478
 
 
 ## 0.7.2